Conflicts:
	app/controllers/user_controller_controller.rb
	app/models/scalarm_authentication.rb

+++ app/models/infrastructure_facades/infrastructure_facade.rb
@@ -16,6 +16,7 @@ require 'mongo_lock'
 # - add_credentials(user, params, session) -> credentials record [MongoActiveRecord] - save credentials to database
 #  -- all params keys are converted to symbols and values are stripped
 # - remove_credentials(record_id, user_id, params) - remove credentials for this infrastructure (e.g. user credentials)
+# - get_credentials(user_id, params) - show collection of credentials records for this infrastructure
 # - enabled_for_user?(user_id) -> true/false - if user with user_id can use this infrastructure
 #
 # Database support methods:
@@ -43,6 +44,8 @@ require 'mongo_lock'
 #   -- this method should not be used directly
 # - _simulation_manager_before_monitor(record) - executed before monitoring single resource
 # - _simulation_manager_after_monitor(record) - executed after monitoring single resource
+# - destroy_unused_credentials(authentication_mode, user) - destroy infrastructure credentials which are not used anymore
+
 
 
 class InfrastructureFacade
@@ -58,7 +61,14 @@ class InfrastructureFacade
     Rails.logger.debug "Preparing configuration for Simulation Manager with id: #{sm_uuid}"
 
     Dir.chdir('/tmp')
-    FileUtils.cp_r(File.join(Rails.root, 'public', 'scalarm_simulation_manager'), "scalarm_simulation_manager_#{sm_uuid}")
+    # using simulation manager implementation based on application config
+    if Rails.configuration.simulation_manager_version == :go
+      # TODO checking somehow the destination server architecture
+      FileUtils.cp_r(File.join(Rails.root, 'public', 'scalarm_simulation_manager_v2', 'bin', 'linux_amd64'), "scalarm_simulation_manager_#{sm_uuid}")
+    elsif Rails.configuration.simulation_manager_version == :ruby
+      FileUtils.cp_r(File.join(Rails.root, 'public', 'scalarm_simulation_manager'), "scalarm_simulation_manager_#{sm_uuid}")
+    end
+
     # prepare sm configuration
     temp_password = SimulationManagerTempPassword.find_by_sm_uuid(sm_uuid)
     temp_password = SimulationManagerTempPassword.create_new_password_for(sm_uuid, experiment_id) if temp_password.nil?
@@ -79,9 +89,9 @@ class InfrastructureFacade
       sm_config['information_service_url'] = Rails.application.secrets.sm_information_service_url
     end
 
-    Rails.logger.debug("Development mode set ? : #{Rails.application.secrets.include?(:information_service_development)}")
+    Rails.logger.debug("Development mode set ? : #{!!Rails.application.secrets.information_service_development}")
 
-    if Rails.application.secrets.include?(:information_service_development)
+    if !!Rails.application.secrets.information_service_development
       sm_config['development'] = true
     end
 
@@ -109,11 +119,16 @@ class InfrastructureFacade
     end
 
     sm_config = {
-        InformationServiceAddress: Rails.application.secrets.information_service_url,
+        InformationServiceAddress: (Rails.application.secrets.sm_information_service_url or
+            Rails.application.secrets.information_service_url),
         Login: temp_password.sm_uuid,
         Password: temp_password.password,
-        Infrastructures: [ infrastructure_name ]
+        Infrastructures: [ infrastructure_name ],
     }
+
+    if Rails.application.secrets.include? :certificate_path
+      sm_config[:ScalarmCertificatePath] = '~/.scalarm_certificate'
+    end
 
     if Rails.application.secrets.include?(:sm_information_service_url)
       sm_config[:InformationServiceAddress] = Rails.application.secrets.sm_information_service_url
@@ -246,12 +261,18 @@ class InfrastructureFacade
     get_sm_records(user_id, experiment_id, attributes).count
   end
 
+  def other_params_for_booster(user_id)
+    {}
+  end
+
+  def get_credentials(*args)
+    raise NotImplementedError
+  end
+
   # -- SimulationManger delegation default implementation --
 
   def _simulation_manager_before_monitor(record); end
   def _simulation_manager_after_monitor(record); end
-
-  private
 
   def create_simulation_manager(record)
     SimulationManager.new(record, self)
@@ -263,4 +284,8 @@ class InfrastructureFacade
   def self.monitoring_package_dir(sm_uuid)
     "scalarm_monitoring_#{sm_uuid}"
   end
+
+  def destroy_unused_credentials(authentication_mode, user); end
+
+  private :create_simulation_manager, :init_resources, :clean_up_resources
 end
